backend/requirements.txt
Django>=5.2.3
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
Pillow>=10.0.0
python-decouple>=3.8
django-extensions>=3.2.3
psycopg2-binary>=2.9.7
django-filter>=23.3
drf-spectacular>=0.26.5 